requirements.txt
Django>=4.2.7,<4.3
mysqlclient>=2.2.0,<2.3
python-dotenv>=1.0.0,<1.1
social-auth-app-django>=5.4.0,<5.5
social-auth-core>=4.5.0,<4.6
Pillow>=10.1.0,<10.2
mysql-connector-python>=8.2.0,<8.3
